Moved template classes definition to cpp files
